Remove misleading URL for win2012r2
When trying to use this file I thought that the URL given might be one that will work for getting a win2012r2 image. I hadn't looked at any other vagrantfiles to notice that it was just a stock URL that wouldn't work for a 2012 box :(

This changes it to make it clearer to anyone reading that they will need to find their own basebox.